tradehut_search/embedding_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch==2.3.0
torchvision==0.18.0
transformers==4.41.0
Pillow==10.3.0
numpy==1.26.4
pydantic==2.7.1
python-multipart==0.0.9
